Npm + code cleanup
