+++ package.json
@@ -12,19 +12,18 @@
     "axios": "^0.18.0",
     "body-parser": "^1.18.3",
     "cron": "^1.6.0",
-    "dotenv": "^5.0.0",
-    "express": "^4.16.2",
+    "dotenv": "^6.2.0",
+    "express": "^4.16.4",
     "lodash.set": "^4.3.2",
-    "mandrill-nodemailer-transport": "^1.0.1",
-    "moment": "^2.22.1",
-    "mongodb": "^3.0.4",
-    "nodemailer": "^4.6.2",
+    "moment": "^2.24.0",
+    "mongodb": "^3.1.13",
+    "nodemailer": "^5.1.1",
     "nodemailer-mandrill-transport": "^1.2.0",
     "request": "^2.88.0",
-    "request-promise": "^4.2.2",
+    "request-promise": "^4.2.4",
     "rtrim": "^1.0.0",
     "strip-hex-prefix": "^1.0.0",
-    "validator": "^9.4.1"
+    "validator": "^10.11.0"
   },
   "scripts": {
     "build-ts": "tsc --declaration",
@@ -45,7 +44,7 @@
     "watch-ts": "tsc -w"
   },
   "devDependencies": {
-    "@types/amqplib": "^0.5.9",
+    "@types/amqplib": "^0.5.11",
     "@types/async": "^2.4.1",
     "@types/async-lock": "^1.1.0",
     "@types/async-retry": "^1.2.1",
@@ -60,18 +59,18 @@
     "@types/hapi": "^18.0.0",
     "@types/hapi-auth-jwt2": "^8.0.2",
     "@types/inert": "^5.1.2",
-    "@types/jest": "^24.0.6",
-    "@types/jsonwebtoken": "^8.3.0",
-    "@types/lodash": "^4.14.120",
-    "@types/lodash.get": "^4.4.4",
-    "@types/lodash.keyby": "^4.6.4",
-    "@types/lodash.set": "^4.3.5",
-    "@types/lru-cache": "^4.1.1",
+    "@types/jest": "^24.0.9",
+    "@types/jsonwebtoken": "^8.3.1",
+    "@types/lodash": "^4.14.122",
+    "@types/lodash.get": "^4.4.6",
+    "@types/lodash.keyby": "^4.6.6",
+    "@types/lodash.set": "^4.3.6",
+    "@types/lru-cache": "^5.1.0",
     "@types/minimist": "^1.2.0",
     "@types/mock-aws-s3": "^2.6.1",
-    "@types/mongodb": "^3.1.19",
+    "@types/mongodb": "^3.1.21",
     "@types/newrelic": "^4.11.0",
-    "@types/node": "^11.9.4",
+    "@types/node": "^11.10.5",
     "@types/nodemailer": "^4.6.6",
     "@types/pdfkit": "^0.7.36",
     "@types/qr-image": "^3.2.1",
@@ -79,19 +78,19 @@
     "@types/request-promise": "^4.1.42",
     "@types/sleep": "0.0.7",
     "@types/validator": "^10.9.0",
-    "@types/vision": "5.3.5",
-    "jest": "^24.1.0",
+    "@types/vision": "^5.3.6",
+    "jest": "^24.3.0",
     "nodemon": "^1.18.10",
     "prettier": "^1.16.4",
     "semantic-release": "^15.13.3",
     "travis-deploy-once": "^5.0.11",
     "ts-jest": "^24.0.0",
-    "ts-node": "^8.0.2",
+    "ts-node": "^8.0.3",
     "ts-node-dev": "^1.0.0-pre.32",
-    "tslint": "^5.12.1",
+    "tslint": "^5.13.1",
     "tslint-config-prettier": "^1.18.0",
-    "tslint-consistent-codestyle": "^1.15.0",
-    "typescript": "^3.3.3",
+    "tslint-consistent-codestyle": "^1.15.1",
+    "typescript": "^3.3.3333",
     "uglify-js": "^3.4.9",
     "uglifyjs-folder": "^1.5.1"
   }
