+++ src/mail/mailer.ts
@@ -4,7 +4,7 @@ import {
 } from '@sendgrid/mail';
 
 import { AttachmentData } from '@sendgrid/helpers/classes/attachment';
-import { ITemplateData } from '../types';
+import { ITemplateData } from '../types/types';
 
 export async function sendUpdate(
   from: string,
@@ -49,47 +49,3 @@ export async function sendPurchased(
     throw error;
   }
 }
-
-// let transporter: nodemailer.Transporter;
-// async function createTransporter() {
-//   return nodemailer.createTransport(
-//     new MandrillTransport({
-//       apiKey: process.env.MANDRILL_API_KEY
-//     })
-//   );
-// }
-
-// export async function send(
-//   emailFrom: string,
-//   emailTo: string,
-//   subject: string,
-//   attachments: Attachment[],
-//   globalMergeVars: IGlobalMergeVar[],
-//   mergeVars: IMergeVar[],
-//   templateSlug: string
-// ) {
-//   if (typeof transporter === 'undefined') {
-//     transporter = await createTransporter();
-//   }
-
-//   const options = {
-//     from: emailFrom,
-//     to: emailTo,
-//     subject,
-//     mandrillOptions: {
-//       template_name: templateSlug,
-//       message: {
-//         global_merge_vars: globalMergeVars,
-//         merge_vars: mergeVars,
-//         attachments
-//       }
-//     }
-//   };
-
-//   transporter.sendMail(options, (error: Error, info) => {
-//     if (error) {
-//       return Promise.reject(error);
-//     }
-//     return Promise.resolve(info);
-//   });
-// }
