+++ src/mail/registries.ts
@@ -1,4 +1,3 @@
-import { getSensorAddressesForSensorId } from '../dapi/sensorRegistry';
 import { getCollection } from '../services/mongo/client';
 
 export async function subscribe(email: string, sensorid: string) {
