+++ src/mail/unsubscribe.ts
@@ -7,10 +7,8 @@ import {
 import { unsubscribe } from './registries';
 
 export async function unsubscribeRoute(req: Request, res: Response) {
-  console.log('Unsubscribe called!');
   const hash = Buffer.from(req.query.hash, 'base64').toString('utf8');
   const parts = hash.split(DELIMITER_HASH);
-  console.log(parts, hash);
   if (parts.length < 1) {
     res.sendStatus(400);
   }
@@ -22,10 +20,9 @@ export async function unsubscribeRoute(req: Request, res: Response) {
     res.sendStatus(400);
   }
   try {
-    console.log(email, sensorid);
     await unsubscribe(email, sensorid);
     const unsubscribedUrl = `${DATABROKER_DAPP_BASE_URL}/unsubscribed`;
-    // res.redirect(unsubscribedUrl);
+    res.redirect(unsubscribedUrl);
   } catch (error) {
     res.send(error).status(200);
   }
