--- src/types/nodemailer-mandrill-transport.d.ts
@@ -1,1 +1,0 @@
-declare module 'nodemailer-mandrill-transport';
