--- src/types/settlemint__lib-crypto.d.ts
@@ -1,1 +1,0 @@
-declare module '@settlemint/lib-crypto';
